+++ R/data_read.R
@@ -292,5 +292,24 @@ data_read <- function(path,
     insight::format_alert("Reading data...")
   }
   out <- rio::import(file = path, ...)
+
+  # for "unknown" data formats (like .RDS), which still can be imported via
+  # "rio::import()", we must check whether we actually have a data frame or
+  # not. Else, tell user.
+  if (!is.data.frame(out)) {
+    tmp <- tryCatch(as.data.frame(out, stringsAsFactors = FALSE), error = function(e) NULL)
+    if (!is.null(tmp)) {
+      out <- tmp
+    } else {
+      if (verbose) {
+        insight::format_warning(
+          paste0("Imported file is no data frame, but of class \"", class(out)[1], "\"."),
+          "Returning file as is. Please check if importing this file was intented."
+        )
+      }
+      return(out)
+    }
+  }
+
   .post_process_imported_data(out, convert_factors, verbose)
 }