+++ tests/testthat/test-standardize_models.R
@@ -14,7 +14,7 @@ test_that("standardize, mlm", {
   m2 <- lm(scale(cbind(mpg, hp)) ~ scale(cyl) + scale(am), data = mtcars)
 
   mz <- standardize(m)
-  expect_identical(coef(mz), coef(m2), ignore_attr = TRUE)
+  expect_equal(coef(mz), coef(m2), ignore_attr = TRUE, tolerance = 1e-4)
 })
 
 test_that("standardize | errors", {