+++ NEWS.md
@@ -1,6 +1,6 @@
-# datawizard (devel)
+# datawizard 0.7.1
 
-BREAKING
+BREAKING CHANGES
 
 * `add_labs()` was renamed into `assign_labels()`. Since `add_labs()` existed
   only for a few days, there will be no alias for backwards compatibility.